test-requirements.txt
molecule==3.1.3
molecule-docker==0.2.4
docker
ansible-lint
pytest-testinfra
flake8
jmespath
